Implemented App Route in local module and fixed warnings
