+++ login/next.config.js
@@ -1,4 +1,6 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {
+    // output: 'export',
+}
 
 module.exports = nextConfig
